starting update sound method, create sound_folder table, starting refactoring sound entity

+++ src/api/sound/soundService.ts
@@ -9,9 +9,12 @@ import { randomString } from "../../utils/utils.js";
 import { StorageError } from "../../errors/general/general.js";
 import { Sound } from "@prisma/client";
 import { AuthorizationError } from "../../errors/auth/auth.js";
+import { getFolderAndFileNameFromPath } from "./helper.js";
 
 type SoundRequestData = {
   folder?: string;
+  name?: string;
+  userId: number;
 };
 interface soundWithUrl extends Sound {
   soundUrl: string;
@@ -46,7 +49,7 @@ export default class SoundService {
       throw new EntityNotFoundError(`Sound with ${id} id not found`, {});
     return sound;
   }
-  async deleteSoundByIdOrError(
+  async getSoundIfUserIsAuthOrError(
     id: number | undefined,
     userId: number | undefined
   ) {
@@ -57,13 +60,20 @@ export default class SoundService {
         404
       );
     //get the audio attempting to delete
-    const soundTodelete = await this.getOneByIdOrError(id);
-    if (soundTodelete.userId !== userId) {
+    const soundToCheck = await this.getOneByIdOrError(id);
+    if (soundToCheck.userId !== userId) {
       throw new AuthorizationError(
-        "This is user is not authorized to delete this sound",
+        "This is user is not authorized to do this sound operation",
         {}
       );
     }
+    return soundToCheck;
+  }
+  async deleteSoundByIdOrError(
+    id: number | undefined,
+    userId: number | undefined
+  ) {
+    const soundTodelete = await this.getSoundIfUserIsAuthOrError(id, userId);
     //delete audio from S3
     if (soundTodelete.path) {
       const resultAudioDeleted = await this.deleteAudioFileOrError(
@@ -71,7 +81,7 @@ export default class SoundService {
       );
       console.log("RESULT OF DELETING SOUND:", resultAudioDeleted);
     }
-
+    if (!id) throw new EntityNotFoundError("Sound id not provided", {});
     const result = await this.SoundRepo.deleteById(id);
 
     if (!result) {
@@ -84,6 +94,50 @@ export default class SoundService {
     return result;
   }
   async createOrError() {}
+  async updateOrError(
+    id: number,
+    data: SoundRequestData,
+    soundFile: Express.Multer.File | undefined
+  ) {
+    const soundToUpdate = await this.getSoundIfUserIsAuthOrError(
+      id,
+      data.userId
+    );
+    const { folder, fileName } = getFolderAndFileNameFromPath(
+      soundToUpdate.path
+    );
+
+    let path = `user-${soundToUpdate.userId}/sounds/${folder}/${fileName}`;
+
+    if (soundFile || data.folder) {
+      if (data.folder && !soundFile) {
+        //move the file to the new folder
+        path = `user-${soundToUpdate.userId}/sounds/${data.folder}/${fileName}`;
+        const result = await this.storage.move(soundToUpdate.path, path);
+      } else if (soundFile) {
+        const resultAudioDeleted = await this.deleteAudioFileOrError(
+          soundToUpdate?.path
+        );
+        path = data.folder
+          ? `user-${soundToUpdate.userId}/sounds/${
+              data.folder
+            }/${randomString()}`
+          : `user-${soundToUpdate.userId}/sounds/${folder}/${randomString()}`;
+        const s3File: S3File = {
+          key: path,
+          body: soundFile.buffer,
+          contentType: soundFile.mimetype,
+        };
+
+        const resultAudioStored = await this.storeAudioFileOrError(s3File);
+      }
+    }
+
+    const soundUpdated = await this.SoundRepo.update(id, {
+      name: data.name || soundFile?.originalname || soundToUpdate.name,
+      path,
+    });
+  }
 
   async getAudioFileUrl(sound: string) {
     try {
@@ -127,21 +181,21 @@ export default class SoundService {
 
   async createManyOrError(
     files: Express.Multer.File[],
-    userId: number | undefined,
     requestData: SoundRequestData
   ) {
     const folderS3 = requestData.folder || "no-categorized";
-    if (!userId) throw new EntityNotFoundError("User not found", {});
-    const soundsToCreate: soundToCreate[] = await Promise.all(
+     const soundsToCreate: soundToCreate[] = await Promise.all(
       files.map(async (file) => {
-        const key = `user-${userId}/sounds/${folderS3}/${randomString()}`;
+        const key = `user-${
+          requestData.userId
+        }/sounds/${folderS3}/${randomString()}`;
         await this.storeAudioFileOrError({
           key,
           body: file.buffer,
           contentType: file.mimetype,
         });
         return {
-          userId,
+          userId: requestData.userId,
           name: path.parse(file.originalname).name,
           path: key,
         };
